+++ dflib/src/main/java/com/nhl/dflib/Scalar.java
@@ -1,6 +1,5 @@
 package com.nhl.dflib;
 
-import com.nhl.dflib.aggregate.ColumnAggregator;
 import com.nhl.dflib.join.JoinPredicate;
 import com.nhl.dflib.row.RowBuilder;
 import com.nhl.dflib.row.RowProxy;