+++ dflib/src/test/java/com/nhl/dflib/HasherTest.java
@@ -1,8 +1,5 @@
-package com.nhl.dflib.map;
+package com.nhl.dflib;
 
-import com.nhl.dflib.DataFrame;
-import com.nhl.dflib.Hasher;
-import com.nhl.dflib.Index;
 import com.nhl.dflib.unit.DFAsserts;
 import org.junit.Test;