+++ RELEASE-NOTES.md
@@ -20,6 +20,7 @@
 * #40 "Series.unique" - a method to produce Series with non-repeating values
 * #43 Column converter to enums for categorical data 
 * #45 DataFrame.renameColumns(UnaryOperator<String>) to upper/lower case column names
+* #46 Move interfaces with public static methods for creating lambdas to the root package
 
 ## Release 0.5